+++ pkg/action/uploader/chart_uploader.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"helm.sh/helm/v3/internal/experimental/pusher"
+	"helm.sh/helm/v3/pkg/action/pusher"
 	"helm.sh/helm/v3/pkg/registry"
 )
